Use higher precision for feature vectors

+++ include/pgmlink/features/featurestore.h
@@ -19,7 +19,7 @@ class Traxel;
 // feature data structures
 //
 
-typedef float feature_type;
+typedef double feature_type;
 typedef std::vector<feature_type> feature_array;
 typedef std::vector<feature_array> feature_arrays;
 typedef std::map<std::string, feature_array> FeatureMap;
